% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{extract_legend}
\alias{extract_legend}
\title{Extract the legend from a ggplot.}
\usage{
extract_legend(gplot)
}
\arguments{
\item{gplot}{A ggplot object.}
}
\description{
Extract the legend from a ggplot.
}

