% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{filter_tidy_data}
\alias{filter_tidy_data}
\title{Filter the tidy data.}
\usage{
filter_tidy_data(tidyData, event_types, date_range, variable)
}
\arguments{
\item{tidyData}{The result of calling tidy_notification_click_data.}

\item{event_types}{A character vector listing the event types to include in
the final chart.}

\item{date_range}{A length-two Date vector containing the min and max date
to include in the filtered dataset.}

\item{variable}{A length-one character vector specifying the variable whose
value will be displayed in the final chart. Can be set to "click",
"unsubscribe", or "manage".}
}
\description{
Filter the tidy data.
}

