% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{generate_final_plot}
\alias{generate_final_plot}
\title{Plot the postprocessed data.}
\usage{
generate_final_plot(processedData)
}
\arguments{
\item{processedData}{The result of calling postprocess.}
}
\description{
Plot the postprocessed data.
}

