% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{get_notification_click_data}
\alias{get_notification_click_data}
\title{Function that runs the notification_settings_query and returns the results
as a data.frame.}
\usage{
get_notification_click_data(con)
}
\arguments{
\item{con}{A database connection object.}
}
\description{
Function that runs the notification_settings_query and returns the results
as a data.frame.
}

