% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\docType{data}
\name{notification_click_query}
\alias{notification_click_query}
\title{SQL query that fetches the notification settings data from Redshift.}
\format{A length-one character vector}
\usage{
notification_click_query
}
\description{
A string containing a SQL query that fetches the notification settings data
from Redshift.
}
\keyword{datasets}

