% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{postprocess}
\alias{postprocess}
\title{Postprocess the tidy data for plotting the final line chart.}
\usage{
postprocess(filteredData, aggregate)
}
\arguments{
\item{filteredData}{The result of calling filter_tidy_data.}

\item{aggregate}{Logical. Should the results for separate event types be
plotted individually (F), or in aggregate (T)?}
}
\description{
Postprocess the tidy data for plotting the final line chart.
}

