% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{tidy_notification_click_data}
\alias{tidy_notification_click_data}
\title{Tidy the data obtained from SQL.}
\usage{
tidy_notification_click_data(clickData)
}
\arguments{
\item{clickData}{The result of calling get_notification_click_data().}
}
\description{
Tidy the data obtained from SQL.
}

