% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\docType{data}
\name{tidy_results}
\alias{tidy_results}
\title{Tidied results}
\format{A data.frame}
\usage{
tidy_results
}
\description{
Contains the most recent tidied results of
tidy_notification_click_data.
}
\keyword{datasets}

